permisos terminados en vista de proveedores

+++ static/js/compras/index.js
@@ -1,6 +1,6 @@
 /* Seccion Proveedor */
 
-function list_proveedor(url_list, url_edit, url_del) {
+function list_proveedor(url_list, url_edit, url_del, perm_change, perm_delete) {
     $('#provTable').DataTable({
         responsive: true,
         pageLength: 6,
@@ -23,8 +23,13 @@ function list_proveedor(url_list, url_edit, url_del) {
             targets: [-1],
             orderable: false,
             render: function (data, type, row) {
-                var buttons = '<button onclick="abrir_modal(\'' + url_edit + row.id + '/\')" class="btn btn-warning btn-sm" title="Editar"><i class="fas fa-edit"></i></button> ';
-                buttons += '<button onclick="delete_ajax_prov(\'' + url_del + row.id + '/\')" title="Borrar" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i></button>';
+                var buttons = ''
+                if(perm_change){
+                    buttons += '<button onclick="abrir_modal(\'' + url_edit + row.id + '/\')" class="btn btn-warning btn-sm" title="Editar"><i class="fas fa-edit"></i></button> ';
+                }
+                if(perm_delete){
+                    buttons += '<button onclick="delete_ajax_prov(\'' + url_del + row.id + '/\')" title="Borrar" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i></button>';
+                }
                 return buttons;
             }
         }]
